Extract `printArrowFunction` to arrow-function.js
